Move collision logic into seperate folder, all tests still passing
