+++ app/lib/collision/array-contains-object.js
