+++ app/lib/collision/index.js
@@ -1,0 +1,3 @@
+var reportCollisions = require('./report-collisions');
+
+module.exports = reportCollisions;
