+++ app/lib/collision/is-colliding.js
